Upgrade to ADCore R3-0
Change the build files to use the ADCore R3-0 build files structure.

Fix the plugin constructor arguments to the super constructor by
removing numParams and adding maxThreads.  (These ADCore changes were
introduced in R3-0 in commit 3af87bd.)

Fix the super call in processCallbacks to call NDPluginDriver's
beginProcessCallbacks method since processCallbacks in NDPluginDriver
was renamed to beginProcessCallbacks.  (This ADCore change was
introduced in R3-0 in commit 3af87bd.)

Call start() in adPythonPluginConfigure, and rename the start variable
in processCallbacks to startTime to avoid shadowing the start method.
(The start method was added in ADCore in R2-5 in commit b1a2f13.)