Add win32 support
Make adPython work on Windows.

Change how adPython calls into the embedded Python to use
PyGILState_Ensure() and PyGILState_Release() which is appropriate for
calling from non-Python-created threads which is the case for adPython
which is called from areaDetector-created threads.  Without this change,
adPython hangs on Windows.  The Python calling idiom is documented in
the following:

* https://docs.python.org/2/c-api/init.html#non-python-created-threads
* https://stackoverflow.com/a/4975906
* https://bugs.python.org/issue1720250#msg57619

Install the plugin scripts so that adPython can reference an installed
pathname rather than a source pathname.

Remove the automatic determination of the Python version, include
pathnames, library pathnames, etc. by executing external commands from
Make since it did not work on Windows.  Replace that with explicit
definitions in configure/CONFIG_SITE.